feat: DraftPageのPC/SP完全統合とコンポーネント共通化
## 主要変更
- PC版・SP版を1つのレスポンシブコンポーネントに統合
- コンポーネント分割による共通化とDRY原則の徹底
- ChatMessageListをPC版ボックススタイルに統一

## 新規作成ファイル
- `DraftPage/index.tsx`: レスポンシブ統合コンポーネント
- `PastDraftResults/`: PC(テーブル)・SP(アコーディオン)対応
- `CurrentRoundStatus/`: 現在ラウンド状況表示の共通化
- `ChatMessageList/`: ボックススタイル統一
- `ChatInputForm/`: 入力フォーム統一
- `useDraftLogic.ts`: 状態管理・ハンドラー共通化

## 削除ファイル
- `pc/index.tsx`: PC版個別コンポーネント
- `sp/index.tsx`: SP版個別コンポーネント

## 技術改善
- 80%のコード重複削除
- TypeScriptエラー0件
- Storybook枠線削除・パターン簡素化
- レスポンシブ対応: useBreakpointValue活用
