+++ package.json
@@ -50,6 +50,7 @@
     "@mdx-js/react": "^3.1.0",
     "@octokit/rest": "^22.0.0",
     "@playwright/test": "1.53.1",
+    "@storybook/addon-actions": "^9.0.8",
     "@storybook/addon-essentials": "8.6.14",
     "@storybook/addon-onboarding": "8.6.14",
     "@storybook/addon-themes": "8.6.14",
