+++ src/components/features/draft/DraftPage/ChatMessageList/index.stories.tsx
@@ -1,0 +1,37 @@
+import type { Meta, StoryObj } from '@storybook/react';
+import { mockChatMessages } from '../mockData';
+import { ChatMessageList } from './index';
+
+const meta: Meta<typeof ChatMessageList> = {
+  title: 'Features/DraftPage/ChatMessageList',
+  component: ChatMessageList,
+  parameters: {
+    layout: 'centered',
+  },
+  args: {
+    messages: mockChatMessages,
+  },
+};
+
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+/**
+ * デフォルト表示
+ * 統一されたボックススタイル
+ */
+export const Default: Story = {
+  args: {
+    variant: 'sp',
+  },
+};
+
+/**
+ * 空のメッセージリスト
+ */
+export const Empty: Story = {
+  args: {
+    variant: 'sp',
+    messages: [],
+  },
+};
