--- src/components/features/draft/DraftPage/pc/index.test.tsx
@@ -1,8 +1,0 @@
-import { DraftPagePC } from './index';
-
-describe('DraftPagePC', () => {
-  test('コンポーネントが正常に定義されている', () => {
-    expect(DraftPagePC).toBeDefined();
-    expect(typeof DraftPagePC).toBe('function');
-  });
-});
