--- src/components/features/draft/DraftPage/sp/index.test.tsx
@@ -1,8 +1,0 @@
-import { DraftPageSp } from './index';
-
-describe('DraftPageSp', () => {
-  test('コンポーネントが正常に定義されている', () => {
-    expect(DraftPageSp).toBeDefined();
-    expect(typeof DraftPageSp).toBe('function');
-  });
-});
